services/crawler/requirements.txt
python-dotenv==0.13.0
Pillow==7.2.0
aiohttp==3.6.2